% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{activity_summary}
\alias{activity_summary}
\title{activity summary for ARGOS datasets}
\usage{
activity_summary(x, trending_N_days = 10)
}
\description{
activity summary for ARGOS datasets
}
\note{
x should contain locationDate (POSIXct) and individual
}
