% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{DBq}
\alias{DBq}
\title{DB query
DB query using default parameters.}
\usage{
DBq(
  x,
  host = getOption("mapps.host"),
  user = getOption("mapps.user"),
  password = getOption("mapps.pwd"),
  db = getOption("mapps.db")
)
}
\value{
a data.table (also for errors). When in shiny query errors are captured by \code{\link[shinytoastr:toastr_success]{shinytoastr::toastr_error()}}
}
\description{
DB query
DB query using default parameters.
}
\examples{
DBq("SELECT 1")
DBq("SELECT err")

}
