% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVER.R
\name{mappServer}
\alias{mappServer}
\title{Shiny server function}
\usage{
mappServer(input, output, session)
}
\arguments{
\item{input}{Shiny server}

\item{output}{Shiny server}

\item{session}{Shiny server}
}
\description{
Shiny server function
}
\note{
hardwired objects expected to be defined in global.R: dbtable, tiles, logo_base64, URL, md.csv
}
