% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{mapp_data}
\alias{mapp_data}
\title{UI interface to database}
\usage{
mapp_data(input)
}
\arguments{
\item{input}{a list containing "mindate", "tagIDs", "locationClass" and LocationDate".
the input is usually defined within a shiny UI like \code{\link[=mappUI]{mappUI()}}.}
}
\description{
UI interface to database
}
